+++ notifications.json
@@ -1,5 +1,12 @@
 [
     {
+        "id": "18",
+        "title": "[UPDATE] Account Mismatch Error",
+        "message": "You can enable Steam Client Beta Participation again. Click to read more",
+        "url": "https://github.com/zevnda/steam-game-idler/discussions/132",
+        "timestamp": "1742540247"
+    },
+    {
         "id": "17",
         "title": "Account Mismatch Error",
         "message": "The Steam Client Beta Participation feature is causing this error. Click here to see how to fix it",
@@ -61,12 +68,5 @@
         "message": "New beta version available",
         "url": "https://github.com/zevnda/steam-game-idler/releases/tag/1.7.0-beta.4",
         "timestamp": "1737923379"
-    },
-    {
-        "id": "8",
-        "title": "Beta v1.7.0-3",
-        "message": "New beta version available",
-        "url": "https://github.com/zevnda/steam-game-idler/releases/tag/1.7.0-beta.3",
-        "timestamp": "1737917746"
     }
 ]
