+++ src/components/customlists/ManualAdd.jsx
@@ -1,16 +1,30 @@
-import { Button, Modal, ModalContent, ModalBody, ModalFooter, ModalHeader, useDisclosure, NumberInput } from '@heroui/react';
+import { Button, Modal, ModalContent, ModalBody, ModalFooter, ModalHeader, useDisclosure, NumberInput, Input } from '@heroui/react';
 import { TbPlus } from 'react-icons/tb';
 
 import useManualAdd from '@/hooks/customlists/useManualAdd';
 
 export default function ManualAdd({ listName, setList }) {
     const { isOpen, onOpen, onOpenChange } = useDisclosure();
-    const { inputValue, isLoading, setInputValue, handleAdd, handleChange } = useManualAdd(listName, setList);
+    const {
+        isLoading,
+        appNameValue,
+        appIdValue,
+        setAppNameValue,
+        setAppIdValue,
+        handleNameChange,
+        handleIdChange,
+        handleAdd
+    } = useManualAdd(listName, setList);
 
     const handleKeyPress = (e, onClose) => {
         if (e.key === 'Enter') {
             handleAdd(onClose);
         }
+    };
+
+    const handleClose = () => {
+        setAppNameValue('');
+        setAppIdValue('');
     };
 
     return (
@@ -23,7 +37,7 @@ export default function ManualAdd({ listName, setList }) {
                 onPress={onOpen}
             />
 
-            <Modal isOpen={isOpen} onOpenChange={onOpenChange} onClose={() => setInputValue('')} className='bg-modalbody text-content' classNames={{ closeButton: ['text-altwhite hover:bg-titlehover duration-200'] }}>
+            <Modal isOpen={isOpen} onOpenChange={onOpenChange} onClose={handleClose} className='bg-modalbody text-content' classNames={{ closeButton: ['text-altwhite hover:bg-titlehover duration-200'] }}>
                 <ModalContent>
                     {(onClose) => (
                         <>
@@ -32,21 +46,32 @@ export default function ManualAdd({ listName, setList }) {
                             </ModalHeader>
                             <ModalBody className='my-4'>
                                 <p className='text-sm mb-2'>
-                                    Add games that you do not own, but have in your library, such as family shared games.
+                                    Manually add a game to your list by entering the game&apos;s name and appid
                                 </p>
+
+                                <Input
+                                    autoFocus
+                                    size='sm'
+                                    placeholder='Game Name'
+                                    value={appNameValue || ''}
+                                    classNames={{
+                                        inputWrapper: ['bg-input border border-border hover:!bg-titlebar rounded-lg group-data-[focus-within=true]:!bg-titlebar group-data-[focus-visible=true]:ring-transparent group-data-[focus-visible=true]:ring-offset-transparent'],
+                                        input: ['!text-content placeholder:text-altwhite/50'],
+                                    }}
+                                    onChange={handleNameChange}
+                                />
 
                                 <NumberInput
                                     hideStepper
-                                    autoFocus
-                                    value={inputValue || 0}
-                                    label='Enter a game ID'
+                                    label='Game ID'
+                                    value={appIdValue || 0}
                                     formatOptions={{ useGrouping: false }}
                                     aria-label='manual add'
                                     classNames={{
                                         inputWrapper: ['bg-input border border-border hover:!bg-titlebar rounded-lg group-data-[focus-within=true]:!bg-titlebar group-data-[focus-visible=true]:ring-transparent group-data-[focus-visible=true]:ring-offset-transparent'],
                                         input: ['text-sm !text-content placeholder:text-altwhite/50'],
                                     }}
-                                    onChange={handleChange}
+                                    onChange={handleIdChange}
                                     onKeyDown={(e) => handleKeyPress(e, onClose)}
                                 />
                             </ModalBody>
@@ -63,7 +88,7 @@ export default function ManualAdd({ listName, setList }) {
                                 <Button
                                     size='sm'
                                     isLoading={isLoading}
-                                    isDisabled={inputValue.length === 0}
+                                    isDisabled={!appNameValue || !appIdValue}
                                     className='font-semibold rounded-lg bg-dynamic text-button'
                                     onPress={() => handleAdd(onClose)}
                                 >
