+++ src/hooks/customlists/useManualAdd.js
@@ -3,30 +3,21 @@ import { invoke } from '@tauri-apps/api/core';
 import { useState, useContext } from 'react';
 
 import { UserContext } from '@/components/contexts/UserContext';
-import { logEvent } from '@/utils/utils';
+import { logEvent } from '@/utils/global/tasks';
 
 export default function useManualAdd(listName, setList) {
     const { userSummary } = useContext(UserContext);
-    const [inputValue, setInputValue] = useState('');
+    const [appNameValue, setAppNameValue] = useState('');
+    const [appIdValue, setAppIdValue] = useState('');
     const [isLoading, setIsLoading] = useState(false);
 
     const handleAdd = async (onClose) => {
         setIsLoading(true);
         try {
-            const res = await invoke('get_game_details', { appId: inputValue.toString() });
-
-            if (res[inputValue].success === false) {
-                setIsLoading(false);
-                addToast({ description: 'A game with that ID does not exist', color: 'danger' });
-                return;
-            }
-
-            const data = res[inputValue].data;
-            const game = { appid: data.steam_appid, name: data.name };
 
             const response = await invoke('add_game_to_custom_list', {
                 steamId: userSummary.steamId,
-                game: { appid: game.appid, name: game.name },
+                game: { appid: parseInt(appIdValue), name: appNameValue },
                 list: listName
             });
             if (response.error) {
@@ -46,14 +37,24 @@ export default function useManualAdd(listName, setList) {
         }
     };
 
-    const handleChange = (e) => {
+    const handleNameChange = (e) => {
+        try {
+            setAppNameValue(e.target.value || '');
+        } catch (error) {
+            addToast({ description: `Error in (handleChange): ${error?.message || error}`, color: 'danger' });
+            console.error('Error in (handleChange):', error);
+            logEvent(`[Error] in (handleChange): ${error}`);
+        }
+    };
+
+    const handleIdChange = (e) => {
         try {
             // If input value starts with 0, remove it
             // Happens when user copy-pastes the appid into the input
             if (e.target?.value.startsWith('0')) {
-                setInputValue(e.target?.value.slice(1) || 0);
+                setAppIdValue(e.target?.value.slice(1) || 0);
             } else {
-                setInputValue(e.target?.value || 0);
+                setAppIdValue(e.target?.value || 0);
             }
         } catch (error) {
             addToast({ description: `Error in (handleChange): ${error?.message || error}`, color: 'danger' });
@@ -63,10 +64,13 @@ export default function useManualAdd(listName, setList) {
     };
 
     return {
-        inputValue,
         isLoading,
-        setInputValue,
-        handleAdd,
-        handleChange
+        appNameValue,
+        appIdValue,
+        setAppNameValue,
+        setAppIdValue,
+        handleNameChange,
+        handleIdChange,
+        handleAdd
     };
 }
