+++ src/hooks/settings/useClearData.js
@@ -3,7 +3,8 @@ import { useContext } from 'react';
 
 import { UserContext } from '@/components/contexts/UserContext';
 import { handleClearLogs } from '@/hooks/settings/useLogs';
-import { logEvent, preserveKeysAndClearData } from '@/utils/utils';
+import { logEvent } from '@/utils/global/tasks';
+import { preserveKeysAndClearData } from '@/utils/global/tasks';
 
 const useClearData = () => {
     const { setUserSummary } = useContext(UserContext);
