+++ src/hooks/settings/useLogs.js
@@ -1,49 +1,49 @@
 import { addToast } from '@heroui/react';
 import { invoke } from '@tauri-apps/api/core';
+import { appDataDir } from '@tauri-apps/api/path';
+import { readTextFile } from '@tauri-apps/plugin-fs';
 import { useState, useEffect } from 'react';
 
-import { logEvent } from '@/utils/utils';
+import { logEvent } from '@/utils/global/tasks';
 
 export const useLogs = () => {
     const [logs, setLogs] = useState([]);
 
     useEffect(() => {
         const fetchLogs = async () => {
-            if (typeof window !== 'undefined' && window.__TAURI__) {
+            try {
+                const fullLogPath = await appDataDir();
+                const logFilePath = `${fullLogPath}\\log.txt`;
+
+                // Check if log file exists
+                let logContents = '';
                 try {
-                    const fullLogPath = await invoke('get_app_log_dir');
-                    const logFilePath = `${fullLogPath}\\log.txt`;
-
-                    // Check if log file exists
-                    let logContents = '';
+                    logContents = await readTextFile(logFilePath);
+                } catch (fileError) {
+                    // Create log file if not exists
+                    console.error('Error in (fetchLogs) - file had to be created:', fileError);
+                    await logEvent('No log file found so one was created');
+                    // Try to read again
                     try {
-                        logContents = await window.__TAURI__.fs.readTextFile(logFilePath);
-                    } catch (fileError) {
-                        // Create log file if not exists
-                        console.error('Error in (fetchLogs) - file had to be created:', fileError);
-                        await logEvent('No log file found so one was created');
-                        // Try to read again
-                        try {
-                            logContents = await window.__TAURI__.fs.readTextFile(logFilePath);
-                        } catch (retryError) {
-                            // Still failed, set empty logs
-                            console.error('Error in (fetchLogs) - unable to create file:', retryError);
-                            setLogs([]);
-                            return;
-                        }
+                        logContents = await readTextFile(logFilePath);
+                    } catch (retryError) {
+                        // Still failed, set empty logs
+                        console.error('Error in (fetchLogs) - unable to create file:', retryError);
+                        setLogs([]);
+                        return;
                     }
-
-                    // Process log contents
-                    const logEntries = logContents.split('\n').filter(entry => entry.trim() !== '').map(entry => {
-                        const [timestamp, message] = entry.split(' + ');
-                        return { timestamp, message };
-                    });
-                    setLogs(logEntries);
-                } catch (error) {
-                    addToast({ description: `Error in (fetchLogs): ${error?.message || error}`, color: 'danger' });
-                    console.error('Error in (fetchLogs):', error);
-                    logEvent(`[Error] in (fetchLogs): ${error}`);
                 }
+
+                // Process log contents
+                const logEntries = logContents.split('\n').filter(entry => entry.trim() !== '').map(entry => {
+                    const [timestamp, message] = entry.split(' + ');
+                    return { timestamp, message };
+                });
+                setLogs(logEntries);
+            } catch (error) {
+                addToast({ description: `Error in (fetchLogs): ${error?.message || error}`, color: 'danger' });
+                console.error('Error in (fetchLogs):', error);
+                logEvent(`[Error] in (fetchLogs): ${error}`);
             }
         };
         fetchLogs();
