+++ src/utils/gameslist/gameCardHandler.js
@@ -2,10 +2,11 @@ import { addToast } from '@heroui/react';
 import { invoke } from '@tauri-apps/api/core';
 
 import ErrorToast from '@/components/ui/ErrorToast';
-import { logEvent, startIdle } from '@/utils/utils';
+import { startIdle } from '@/utils/global/idle';
+import { logEvent } from '@/utils/global/tasks';
 
 export const handleIdle = async (item) => {
-    const success = await startIdle(item.appid, item.name, false, true);
+    const success = await startIdle(item.appid, item.name, true);
     if (success) {
         addToast({ description: `Started idling ${item.name} (${item.appid})`, color: 'success' });
     } else {
@@ -17,7 +18,7 @@ export const handleStopIdle = async (item, idleGamesList, setIdleGamesList) => {
     const game = idleGamesList.find((game) => game.appid === item.appid);
     try {
         const response = await invoke('kill_process_by_pid', { pid: game.pid });
-        if (response.includes('success')) {
+        if (response.success) {
             setIdleGamesList(idleGamesList.filter((i) => i.pid !== item.pid));
             addToast({ description: `Stopped idling ${item.name}`, color: 'success' });
         } else {
@@ -32,7 +33,7 @@ export const handleStopIdle = async (item, idleGamesList, setIdleGamesList) => {
 
 export const viewAchievments = async (item, setAppId, setAppName, setShowAchievements) => {
     // Check if Steam is running
-    const steamRunning = await invoke('check_status');
+    const steamRunning = await invoke('is_steam_running');
     if (!steamRunning) {
         return addToast({
             description: <ErrorToast
